boards: renesas: xen: add board for Xen DomD on R-Car Gen4
Add separate board for running U-boot as intermediate bootloader
for Xen driver domain on Renesas Gen4 Spider board. It is needed
to use Renesas drivers for peripherals, that is shared with DomD.
As these driver is depandand on ARCH_RMOBILE and other Renesas
Kconfig, it was needed to merge Xen and Renesas configs.

Signed-off-by: Dmytro Firsov <dmytro_firsov@epam.com>
Acked-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Acked-by: Sergiy Kibrik <Sergiy_Kibrik@epam.com>